+++ server/lib/hook.js
@@ -1,22 +1,9 @@
-import Promise from 'bluebird';
-import request from 'superagent';
+import axios from 'axios';
 
 export default function(report, hook) {
   if (!hook) {
     return Promise.resolve();
   }
 
-  return new Promise((resolve, reject) => {
-    request
-      .post(hook)
-      .send(report)
-      .set('Accept', 'application/json')
-      .end((err) => {
-        if (err) {
-          return reject(err);
-        }
-
-        return resolve();
-      });
-  });
+  return axios.post(hook, report);
 }