+++ server/lib/providers/bitbucketApi.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import request from 'request';
+import axios from 'axios';
 
 export default class Bitbucket {
   constructor(options) {
@@ -20,42 +20,9 @@ export default class Bitbucket {
       }
     }, options);
 
-    this.request = (opts, cb) => {
-      request(opts, (error, response, data) => {
-        if (error) {
-          cb(error);
-        } else {
-          try {
-            data = JSON.parse(data); // eslint-disable-line no-param-reassign
-            if (typeof data.errors !== 'undefined') {
-              cb(data.errors);
-            } else if (response.statusCode !== 200) {
-              cb(this.generateApiError(opts.url, response.statusCode, response));
-            } else {
-              cb(null, data);
-            }
-          } catch (parseError) {
-            if (response.statusCode === 200) {
-              cb(null, data);
-            } else {
-              cb(this.generateApiError(opts.url, response.statusCode, response));
-            }
-          }
-        }
-      }).auth(this.options.user_name, this.options.password, true);
-    };
-
-    this.generateApiError = (url, statusCode, response = null) => {
-      const error = new Error(`Error ${statusCode} when calling GET '${url}' (username: ${this.options.user_name})`);
-      error.status = statusCode;
-      error.statusCode = statusCode;
-      error.report = `status: ${statusCode}
-      user: ${this.options.user_name}
-      url: ${url}
-      method: get
-      response: ${JSON.stringify(response)}`;
-      return error;
-    };
+    this.request = (opts) =>
+      axios({ ...opts, auth: { username: this.options.user_name, password: this.options.password } })
+        .then(response => response.data);
   }
 }
 
@@ -72,42 +39,34 @@ Bitbucket.prototype.buildEndpoint = function buildEndpoint(path, params) {
   };
 };
 
-Bitbucket.prototype.getAll = function getAll(options, callback) {
+Bitbucket.prototype.getAll = function getAll(options) {
   const perPage = 100;
   const result = [];
 
   options.url = options.url + '?pagelen=' + perPage;
 
-  const getPage = (url, cb) => {
+  const getPage = (url) => {
     const pageOptions = { ...options };
     if (url) {
       pageOptions.url = url;
     }
 
-    this.request(pageOptions, (error, data) => {
-      if (error) {
-        return cb(error);
-      }
+    return this.request(pageOptions)
+      .then(data => {
+        data.values.forEach(item => result.push(item));
 
-      data.values.forEach(item => result.push(item));
+        if (data.next) {
+          return getPage(data.next);
+        }
 
-      if (data.next) {
-        return getPage(data.next, cb);
-      }
-
-      return cb(null, result);
-    });
+        return result;
+      });
   };
 
-  return getPage(null, callback);
+  return getPage();
 };
 
-Bitbucket.prototype.doRequest = function doRequest(isTree, path, params, callback) {
-  if (typeof params === 'function') {
-    callback = params; // eslint-disable-line no-param-reassign
-    params = {}; // eslint-disable-line no-param-reassign
-  }
-
+Bitbucket.prototype.doRequest = function doRequest(isTree, path, params) {
   const endpoint = this.buildEndpoint(path, params);
 
   const options = {
@@ -119,28 +78,24 @@ Bitbucket.prototype.doRequest = function doRequest(isTree, path, params, callbac
   options.headers = this.options.request_options.headers;
 
   if (isTree) {
-    return this.getAll(options, (error, data) => {
-      if (error) {
-        callback(error);
-      } else {
-        callback(null, data);
-      }
-    });
+    return this.getAll(options);
   }
 
-  return this.request(options, (error, data) => {
-    if (error) {
-      callback(error);
-    } else {
-      callback(null, data);
-    }
-  });
+  return this.request(options);
 };
 
-Bitbucket.prototype.getTree = function getTree(url, params, callback) {
-  return this.doRequest(true, url, params, callback);
+Bitbucket.prototype.getTree = function getTree(url, params) {
+  return this.doRequest(true, url, params)
+    .then(res => res)
+    .catch((err) => {
+      if (err.response && err.response.status === 404) {
+        return Promise.resolve([]);
+      }
+
+      return Promise.reject(err);
+    });
 };
 
-Bitbucket.prototype.get = function get(url, params, callback) {
-  return this.doRequest(false, url, params, callback);
+Bitbucket.prototype.get = function get(url, params) {
+  return this.doRequest(false, url, params);
 };