+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### auth0-bitbucket-deploy v2.8.0
+### auth0-github-deploy v2.8.0
+### auth0-gitlab-deploy v2.9.0
+### auth0-visualstudio-deploy v2.7.0
+
+- #### Changed
+  - `request-promise` is replaced by `axios`.
+  
 ## [1.2.3] - 2019-03-25
 
 ### auth0-bitbucket-deploy v2.7.2